refactor: use Long for mainTask and subTask ids

+++ app/src/main/java/com/example/taskgenerator/data/entity/Main_task_entity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "main_task")
 data class Main_task_entity(
     @PrimaryKey(autoGenerate = true)
-    val mainTaskId: Int,
+    val mainTaskId: Long,
     var title: String,
     var description: String,
     val taskType: String? =null,
